renamed app (fixtures -> fixture)
